Moved components for calendar to it's own folder
